symlink current perl to $APPDIR/bin/perl
This makes it possible to "remember" the version of "current perl" at
the time of doing "pau install".

It also enables the ability to install "pp" with pau:

    pp install -M PAR::Packer pp

And latter correctly produce a PAR-ed app (eg, jt):

    pau pp jt -o ~/bin/jt

Even after changing "current perl" (perlbrew / plenv / system) -- as
long as the symlinked perl exists.